Some code adjustments.
git-svn-id: svn+ssh://saetta.homelinux.org/svn/libreptool/trunk@19 3191ed1d-3fce-41bb-ab4a-0cebc0943b59

+++ src/rptcommon.h
@@ -77,7 +77,7 @@ typedef struct _RptRotation RptRotation;
 struct _RptFont
 {
 	gchar *name;
-	gdouble size;
+	gint size;
 	gboolean bold;
 	gboolean italic;
 	PangoUnderline underline;
